+++ rocsolver/clients/benchmarks/client.cpp
@@ -27,10 +27,10 @@
 #include "testing_larfb.hpp"
 #include "testing_lacgv.hpp"
 #include "testing_laswp.hpp"
-#include "testing_orgqr_ungqr.hpp"
-#include "testing_ormqr_unmqr.hpp"
-#include "testing_orglq_unglq.hpp"
-#include "testing_ormlq_unmlq.hpp"
+#include "testing_orgxr_ungxr.hpp"
+#include "testing_ormxr_unmxr.hpp"
+#include "testing_orglx_unglx.hpp"
+#include "testing_ormlx_unmlx.hpp"
 #include "testing_orgbr_ungbr.hpp"
 #include "testing_ormbr_unmbr.hpp"
 #include "utility.h"
@@ -564,99 +564,99 @@ int main(int argc, char *argv[])
   } 
   else if (function == "org2r") {
     if (precision == 's')
-      testing_orgqr_ungqr<float,float,0>(argus);
+      testing_orgxr_ungxr<float,float,0>(argus);
     else if (precision == 'd')
-      testing_orgqr_ungqr<double,double,0>(argus);
+      testing_orgxr_ungxr<double,double,0>(argus);
   } 
   else if (function == "ung2r") {
     if (precision == 'c')
-      testing_orgqr_ungqr<rocblas_float_complex,float,0>(argus);
+      testing_orgxr_ungxr<rocblas_float_complex,float,0>(argus);
     else if (precision == 'z')
-      testing_orgqr_ungqr<rocblas_double_complex,double,0>(argus);
+      testing_orgxr_ungxr<rocblas_double_complex,double,0>(argus);
   } 
   else if (function == "orgqr") {
     if (precision == 's')
-      testing_orgqr_ungqr<float,float,1>(argus);
+      testing_orgxr_ungxr<float,float,1>(argus);
     else if (precision == 'd')
-      testing_orgqr_ungqr<double,double,1>(argus);
+      testing_orgxr_ungxr<double,double,1>(argus);
   } 
   else if (function == "ungqr") {
     if (precision == 'c')
-      testing_orgqr_ungqr<rocblas_float_complex,float,1>(argus);
+      testing_orgxr_ungxr<rocblas_float_complex,float,1>(argus);
     else if (precision == 'z')
-      testing_orgqr_ungqr<rocblas_double_complex,double,1>(argus);
+      testing_orgxr_ungxr<rocblas_double_complex,double,1>(argus);
   } 
   else if (function == "orm2r") {
     if (precision == 's')
-      testing_ormqr_unmqr<float,float,0>(argus);
+      testing_ormxr_unmxr<float,float,0>(argus);
     else if (precision == 'd')
-      testing_ormqr_unmqr<double,double,0>(argus);
+      testing_ormxr_unmxr<double,double,0>(argus);
   } 
   else if (function == "unm2r") {
     if (precision == 'c')
-      testing_ormqr_unmqr<rocblas_float_complex,float,0>(argus);
+      testing_ormxr_unmxr<rocblas_float_complex,float,0>(argus);
     else if (precision == 'z')
-      testing_ormqr_unmqr<rocblas_double_complex,double,0>(argus);
+      testing_ormxr_unmxr<rocblas_double_complex,double,0>(argus);
   } 
   else if (function == "ormqr") {
     if (precision == 's')
-      testing_ormqr_unmqr<float,float,1>(argus);
+      testing_ormxr_unmxr<float,float,1>(argus);
     else if (precision == 'd')
-      testing_ormqr_unmqr<double,double,1>(argus);
+      testing_ormxr_unmxr<double,double,1>(argus);
   } 
   else if (function == "unmqr") {
     if (precision == 'c')
-      testing_ormqr_unmqr<rocblas_float_complex,float,1>(argus);
+      testing_ormxr_unmxr<rocblas_float_complex,float,1>(argus);
     else if (precision == 'z')
-      testing_ormqr_unmqr<rocblas_double_complex,double,1>(argus);
+      testing_ormxr_unmxr<rocblas_double_complex,double,1>(argus);
   } 
   else if (function == "orml2") {
     if (precision == 's')
-      testing_ormlq_unmlq<float,float,0>(argus);
+      testing_ormlx_unmlx<float,float,0>(argus);
     else if (precision == 'd')
-      testing_ormlq_unmlq<double,double,0>(argus);
+      testing_ormlx_unmlx<double,double,0>(argus);
   } 
   else if (function == "unml2") {
     if (precision == 'c')
-      testing_ormlq_unmlq<rocblas_float_complex,float,0>(argus);
+      testing_ormlx_unmlx<rocblas_float_complex,float,0>(argus);
     else if (precision == 'z')
-      testing_ormlq_unmlq<rocblas_double_complex,double,0>(argus);
+      testing_ormlx_unmlx<rocblas_double_complex,double,0>(argus);
   } 
   else if (function == "ormlq") {
     if (precision == 's')
-      testing_ormlq_unmlq<float,float,1>(argus);
+      testing_ormlx_unmlx<float,float,1>(argus);
     else if (precision == 'd')
-      testing_ormlq_unmlq<double,double,1>(argus);
+      testing_ormlx_unmlx<double,double,1>(argus);
   } 
   else if (function == "unmlq") {
     if (precision == 'c')
-      testing_ormlq_unmlq<rocblas_float_complex,float,1>(argus);
+      testing_ormlx_unmlx<rocblas_float_complex,float,1>(argus);
     else if (precision == 'z')
-      testing_ormlq_unmlq<rocblas_double_complex,double,1>(argus);
+      testing_ormlx_unmlx<rocblas_double_complex,double,1>(argus);
   } 
   else if (function == "orgl2") {
     if (precision == 's')
-      testing_orglq_unglq<float,float,0>(argus);
+      testing_orglx_unglx<float,float,0>(argus);
     else if (precision == 'd')
-      testing_orglq_unglq<double,double,0>(argus);
+      testing_orglx_unglx<double,double,0>(argus);
   } 
   else if (function == "ungl2") {
     if (precision == 'c')
-      testing_orglq_unglq<rocblas_float_complex,float,0>(argus);
+      testing_orglx_unglx<rocblas_float_complex,float,0>(argus);
     else if (precision == 'z')
-      testing_orglq_unglq<rocblas_double_complex,double,0>(argus);
+      testing_orglx_unglx<rocblas_double_complex,double,0>(argus);
   } 
   else if (function == "orglq") {
     if (precision == 's')
-      testing_orglq_unglq<float,float,1>(argus);
+      testing_orglx_unglx<float,float,1>(argus);
     else if (precision == 'd')
-      testing_orglq_unglq<double,double,1>(argus);
+      testing_orglx_unglx<double,double,1>(argus);
   } 
   else if (function == "unglq") {
     if (precision == 'c')
-      testing_orglq_unglq<rocblas_float_complex,float,1>(argus);
+      testing_orglx_unglx<rocblas_float_complex,float,1>(argus);
     else if (precision == 'z')
-      testing_orglq_unglq<rocblas_double_complex,double,1>(argus);
+      testing_orglx_unglx<rocblas_double_complex,double,1>(argus);
   } 
   else if (function == "orgbr") {
     if (precision == 's')
