+++ rocsolver/clients/gtest/orglx_unglx_gtest.cpp
@@ -3,7 +3,7 @@
  *
  * ************************************************************************ */
 
-#include "testing_orgqr_ungqr.hpp"
+#include "testing_orglx_unglx.hpp"
 #include "utility.h"
 #include <gtest/gtest.h>
 #include <math.h>
@@ -17,37 +17,37 @@ using ::testing::ValuesIn;
 using namespace std;
 
 
-typedef std::tuple<vector<int>, vector<int>> orgqr_tuple;
+typedef std::tuple<vector<int>, int> orglq_tuple;
 
-// vector of vector, each vector is a {M, lda};
+// vector of vector, each vector is a {M, lda, K};
 const vector<vector<int>> m_size_range = {
-    {0, 1}, {-1, 1}, {20, 5}, {50, 50}, {70, 100}, {130, 130}
+    {0, 1, 1}, {-1, 1, 1}, {20, 5, 1}, {10, 10, 20}, {10, 10, 10}, {20, 50, 20}, {130, 130, 50}
 };
 
-// each is a {N, K}
-const vector<vector<int>> n_size_range = {
-    {-1, 1}, {0, 1}, {1, -1}, {10, 0}, {10, 20}, {20, 20}, {130, 130}
+// each is a N
+const vector<int> n_size_range = {
+    -1, 0, 50, 70, 130
 };
 
 const vector<vector<int>> large_m_size_range = {
-    {152, 152}, {640, 640}, {1000, 1024}, {2000, 2000} 
+    {164, 164, 130}, {198, 640, 198}, {130, 130, 130}, {220, 220, 140}, {400, 400, 200} 
 };
 
-const vector<vector<int>> large_n_size_range = {
-    {164, 162}, {198, 140}, {130, 130}, {220, 220}, {400, 200}
+const vector<int> large_n_size_range = {
+    130, 400, 640, 1000, 2000
 };
 
 
-Arguments setup_arguments_org(orgqr_tuple tup) 
+Arguments setup_arguments_orglq(orglq_tuple tup) 
 {
     vector<int> m_size = std::get<0>(tup);
-    vector<int> n_size = std::get<1>(tup);
+    int n_size = std::get<1>(tup);
 
     Arguments arg;
 
     arg.M = m_size[0];
-    arg.N = n_size[0];
-    arg.K = n_size[1];
+    arg.N = n_size;
+    arg.K = m_size[2];
     arg.lda = m_size[1];
 
     arg.timing = 0;
@@ -55,23 +55,23 @@ Arguments setup_arguments_org(orgqr_tuple tup)
     return arg;
 }
 
-class OrthoColGen : public ::TestWithParam<orgqr_tuple> {
+class OrthoRowGen : public ::TestWithParam<orglq_tuple> {
 protected:
-    OrthoColGen() {}
-    virtual ~OrthoColGen() {}
+    OrthoRowGen() {}
+    virtual ~OrthoRowGen() {}
     virtual void SetUp() {}
     virtual void TearDown() {}
 };
 
-TEST_P(OrthoColGen, org2r_float) {
-    Arguments arg = setup_arguments_org(GetParam());
+TEST_P(OrthoRowGen, orgl2_float) {
+    Arguments arg = setup_arguments_orglq(GetParam());
 
-    rocblas_status status = testing_orgqr_ungqr<float,float,0>(arg);
+    rocblas_status status = testing_orglx_unglx<float,float,0>(arg);
 
     // if not success, then the input argument is problematic, so detect the error
     // message
     if (status != rocblas_status_success) {
-        if (arg.M < 0 || arg.N < 0 || arg.K < 0 || arg.N > arg.M || arg.K > arg.N) {
+        if (arg.M < 0 || arg.N < 0 || arg.K < 0 || arg.N < arg.M || arg.K > arg.M) {
             EXPECT_EQ(rocblas_status_invalid_size, status);
         } else if (arg.lda < arg.M) {
             EXPECT_EQ(rocblas_status_invalid_size, status);
@@ -79,15 +79,15 @@ TEST_P(OrthoColGen, org2r_float) {
     }
 }
 
-TEST_P(OrthoColGen, org2r_double) {
-    Arguments arg = setup_arguments_org(GetParam());
+TEST_P(OrthoRowGen, orgl2_double) {
+    Arguments arg = setup_arguments_orglq(GetParam());
 
-    rocblas_status status = testing_orgqr_ungqr<double,double,0>(arg);
+    rocblas_status status = testing_orglx_unglx<double,double,0>(arg);
 
     // if not success, then the input argument is problematic, so detect the error
     // message
     if (status != rocblas_status_success) {
-        if (arg.M < 0 || arg.N < 0 || arg.K < 0 || arg.N > arg.M || arg.K > arg.N) {
+        if (arg.M < 0 || arg.N < 0 || arg.K < 0 || arg.N < arg.M || arg.K > arg.M) {
             EXPECT_EQ(rocblas_status_invalid_size, status);
         } else if (arg.lda < arg.M) {
             EXPECT_EQ(rocblas_status_invalid_size, status);
@@ -95,15 +95,15 @@ TEST_P(OrthoColGen, org2r_double) {
     }
 }
 
-TEST_P(OrthoColGen, ung2r_float_complex) {
-    Arguments arg = setup_arguments_org(GetParam());
+TEST_P(OrthoRowGen, ungl2_float_complex) {
+    Arguments arg = setup_arguments_orglq(GetParam());
 
-    rocblas_status status = testing_orgqr_ungqr<rocblas_float_complex,float,0>(arg);
+    rocblas_status status = testing_orglx_unglx<rocblas_float_complex,float,0>(arg);
 
     // if not success, then the input argument is problematic, so detect the error
     // message
     if (status != rocblas_status_success) {
-        if (arg.M < 0 || arg.N < 0 || arg.K < 0 || arg.N > arg.M || arg.K > arg.N) {
+        if (arg.M < 0 || arg.N < 0 || arg.K < 0 || arg.N < arg.M || arg.K > arg.M) {
             EXPECT_EQ(rocblas_status_invalid_size, status);
         } else if (arg.lda < arg.M) {
             EXPECT_EQ(rocblas_status_invalid_size, status);
@@ -111,15 +111,15 @@ TEST_P(OrthoColGen, ung2r_float_complex) {
     }
 }
 
-TEST_P(OrthoColGen, ung2r_double_complex) {
-    Arguments arg = setup_arguments_org(GetParam());
+TEST_P(OrthoRowGen, ungl2_double_complex) {
+    Arguments arg = setup_arguments_orglq(GetParam());
 
-    rocblas_status status = testing_orgqr_ungqr<rocblas_double_complex,double,0>(arg);
+    rocblas_status status = testing_orglx_unglx<rocblas_double_complex,double,0>(arg);
 
     // if not success, then the input argument is problematic, so detect the error
     // message
     if (status != rocblas_status_success) {
-        if (arg.M < 0 || arg.N < 0 || arg.K < 0 || arg.N > arg.M || arg.K > arg.N) {
+        if (arg.M < 0 || arg.N < 0 || arg.K < 0 || arg.N < arg.M || arg.K > arg.M) {
             EXPECT_EQ(rocblas_status_invalid_size, status);
         } else if (arg.lda < arg.M) {
             EXPECT_EQ(rocblas_status_invalid_size, status);
@@ -127,15 +127,15 @@ TEST_P(OrthoColGen, ung2r_double_complex) {
     }
 }
 
-TEST_P(OrthoColGen, orgqr_float) {
-    Arguments arg = setup_arguments_org(GetParam());
+TEST_P(OrthoRowGen, orglq_float) {
+    Arguments arg = setup_arguments_orglq(GetParam());
 
-    rocblas_status status = testing_orgqr_ungqr<float,float,1>(arg);
+    rocblas_status status = testing_orglx_unglx<float,float,1>(arg);
 
     // if not success, then the input argument is problematic, so detect the error
     // message
     if (status != rocblas_status_success) {
-        if (arg.M < 0 || arg.N < 0 || arg.K < 0 || arg.N > arg.M || arg.K > arg.N) {
+        if (arg.M < 0 || arg.N < 0 || arg.K < 0 || arg.N < arg.M || arg.K > arg.M) {
             EXPECT_EQ(rocblas_status_invalid_size, status);
         } else if (arg.lda < arg.M) {
             EXPECT_EQ(rocblas_status_invalid_size, status);
@@ -143,15 +143,15 @@ TEST_P(OrthoColGen, orgqr_float) {
     }
 }
 
-TEST_P(OrthoColGen, orgqr_double) {
-    Arguments arg = setup_arguments_org(GetParam());
+TEST_P(OrthoRowGen, orglq_double) {
+    Arguments arg = setup_arguments_orglq(GetParam());
 
-    rocblas_status status = testing_orgqr_ungqr<double,double,1>(arg);
+    rocblas_status status = testing_orglx_unglx<double,double,1>(arg);
 
     // if not success, then the input argument is problematic, so detect the error
     // message
     if (status != rocblas_status_success) {
-        if (arg.M < 0 || arg.N < 0 || arg.K < 0 || arg.N > arg.M || arg.K > arg.N) {
+        if (arg.M < 0 || arg.N < 0 || arg.K < 0 || arg.N < arg.M || arg.K > arg.M) {
             EXPECT_EQ(rocblas_status_invalid_size, status);
         } else if (arg.lda < arg.M) {
             EXPECT_EQ(rocblas_status_invalid_size, status);
@@ -159,15 +159,15 @@ TEST_P(OrthoColGen, orgqr_double) {
     }
 }
 
-TEST_P(OrthoColGen, ungqr_float_complex) {
-    Arguments arg = setup_arguments_org(GetParam());
+TEST_P(OrthoRowGen, unglq_float_complex) {
+    Arguments arg = setup_arguments_orglq(GetParam());
 
-    rocblas_status status = testing_orgqr_ungqr<rocblas_float_complex,float,1>(arg);
+    rocblas_status status = testing_orglx_unglx<rocblas_float_complex,float,1>(arg);
 
     // if not success, then the input argument is problematic, so detect the error
     // message
     if (status != rocblas_status_success) {
-        if (arg.M < 0 || arg.N < 0 || arg.K < 0 || arg.N > arg.M || arg.K > arg.N) {
+        if (arg.M < 0 || arg.N < 0 || arg.K < 0 || arg.N < arg.M || arg.K > arg.M) {
             EXPECT_EQ(rocblas_status_invalid_size, status);
         } else if (arg.lda < arg.M) {
             EXPECT_EQ(rocblas_status_invalid_size, status);
@@ -175,15 +175,15 @@ TEST_P(OrthoColGen, ungqr_float_complex) {
     }
 }
 
-TEST_P(OrthoColGen, ungqr_double_complex) {
-    Arguments arg = setup_arguments_org(GetParam());
+TEST_P(OrthoRowGen, unglq_double_complex) {
+    Arguments arg = setup_arguments_orglq(GetParam());
 
-    rocblas_status status = testing_orgqr_ungqr<rocblas_double_complex,double,1>(arg);
+    rocblas_status status = testing_orglx_unglx<rocblas_double_complex,double,1>(arg);
 
     // if not success, then the input argument is problematic, so detect the error
     // message
     if (status != rocblas_status_success) {
-        if (arg.M < 0 || arg.N < 0 || arg.K < 0 || arg.N > arg.M || arg.K > arg.N) {
+        if (arg.M < 0 || arg.N < 0 || arg.K < 0 || arg.N < arg.M || arg.K > arg.M) {
             EXPECT_EQ(rocblas_status_invalid_size, status);
         } else if (arg.lda < arg.M) {
             EXPECT_EQ(rocblas_status_invalid_size, status);
@@ -192,10 +192,10 @@ TEST_P(OrthoColGen, ungqr_double_complex) {
 }
 
 
-INSTANTIATE_TEST_CASE_P(daily_lapack, OrthoColGen,
+INSTANTIATE_TEST_CASE_P(daily_lapack, OrthoRowGen,
                         Combine(ValuesIn(large_m_size_range),
                                 ValuesIn(large_n_size_range)));
 
-INSTANTIATE_TEST_CASE_P(checkin_lapack, OrthoColGen,
+INSTANTIATE_TEST_CASE_P(checkin_lapack, OrthoRowGen,
                         Combine(ValuesIn(m_size_range),
                                 ValuesIn(n_size_range)));
